feat: OpenWeather API

+++ app/src/main/java/com/benrostudios/airtree/models/Coord.kt
@@ -4,6 +4,6 @@ package com.benrostudios.airtree.models
 import com.google.gson.annotations.SerializedName
 
 data class Coord(
-    var lat: Int, // 50
-    var lon: Int // 50
+    var lat: Double, // 50
+    var lon: Double // 50
 )
