+++ app/src/main/java/com/benrostudios/airtree/models/Main.kt
@@ -4,5 +4,5 @@ package com.benrostudios.airtree.models
 import com.google.gson.annotations.SerializedName
 
 data class Main(
-    var aqi: Int // 1
+    var aqi: Double // 1
 )
